% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifesto.R
\name{canno}
\alias{canno}
\title{canno  - process csv manifest into annotation object for illumina methylation preprocessing}
\usage{
canno(man = "EPIC-8v2-0_A1.csv", name = NULL)
}
\arguments{
\item{man}{name}
}
\value{
IlluminaMethylationManifest
}
\description{
canno  - process csv manifest into annotation object for illumina methylation preprocessing
}
\details{
This is based on the scripts shipped with minfi annotation packages. 
It is based on existing 
Illumina Human Methylation csv format manifests, but because 
reverse-engineered, may require
updates to work on future products.
}
