\name{colnames-methods}
\docType{methods}
\alias{colnames-methods}
\alias{colnames,MethyLumiSet-method}
\title{  Methods for Function \code{colnames} in Package \pkg{wateRmelon} }
\description{
   Methods for function \code{colnames} in package \pkg{wateRmelon}.
}
\section{Methods}{
\describe{


\item{\code{signature(x = "MethyLumiSet")}}{

returns the sample names

}
}}
\keyword{methods}
\keyword{ ~~ other possible keyword(s) ~~ }
