% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createAnnotation.R
\name{.createAnnotation}
\alias{.createAnnotation}
\title{Internal function to guess correct pData and retrieve using minfi}
\usage{
.createAnnotation(object)
}
\description{
Internal function to guess correct pData and retrieve using minfi
}
