% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_minfi.R
\name{.getManifestString}
\alias{.getManifestString}
\alias{got}
\alias{fot}
\alias{pop}
\title{Internal functions for Illumina i450 normalization functions}
\usage{
.getManifestString(annotation)
}
\arguments{
\item{x}{a MethyLumiSet}

\item{obj}{a MethylSet}

\item{fd}{a character vector of the desired annotation columns}

\item{rn}{a character vector of the desired features}
}
\description{
got and fot find the annotation column differentiating type I and type II
assays in MethylSet (got) or MethyLumiSet (fot) objects. pop extracts
columns from IlluminaHumanMethylation450k.db
}
\details{
\code{got} returns a character vector of 'I' and 'II', \code{fot} returns
the index of the relevant column. \code{pop} returns a data frame
}
\references{
Pidsley R, Wong CCY, Volta M, Lunnon K, Mill J, Schalkwyk LC: A
data-driven approach to preprocessing Illumina 450K methylation array data
(submitted)
}
\author{
Leonard.Schalkwyk@kcl.ac.uk
}
