% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_bigmelon.R
\name{epicv2clean.default}
\alias{epicv2clean.default}
\alias{epicv2clean.gds.class}
\alias{epicv2clean}
\title{Strip and subset EPICv2 data to work with legacy data and methods}
\usage{
\method{epicv2clean}{default}(x)
}
\description{
Returns an object with rownames stripped of the EPICv2 suffixes, duplicate probes are omitted.
}
\details{
EPICv2 manifests contain a few thousand probes with up to 10 replicate syntheses. 
To accomodate this a modified naming scheme is used, so none of the probe names match
those on the EPIC and previous arrays (even though most of the probes are the same sequence
and presumably similar performance). 

This simple function relies on the rowname and subsetting methods and will work for matrix, 
dataframe, MethyLumiSet, or MethylSet objects, and there is a method for gds (bigmelon) objects.
}
