% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateSex.R
\name{estimateSex}
\alias{estimateSex}
\alias{sex_coef}
\title{Predict sex by using robust sex-related CpG sites on ChrX and ChrY}
\usage{
estimateSex(betas, do_plot = FALSE)
}
\arguments{
\item{betas}{raw beta values, ideally: beta = M / (M + U + 100).}

\item{do_plot}{logical. Should plot the predicted results? Default: FALSE}
}
\value{
dataframe contains predicted sex information.
}
\description{
Predict sex by using robust sex-related CpG sites on ChrX and ChrY
}
\examples{
data(melon)
pred_XY <- estimateSex(betas(melon), do_plot=TRUE)
}
\author{
Wang, Yucheng
}
