\name{genki-methods}
\docType{methods}
\alias{genki-methods}
\alias{genki,exprmethy450-method}
\alias{genki,MethylSet-method}
\alias{genki,MethyLumiSet-method}
\alias{genki,RGChannelSet-method}
\title{Methods for Function \code{genki} in Package \pkg{wateRmelon}}
\description{
Methods for function \code{genki} in package \pkg{wateRmelon}. 
Please see \code{\link{genki}} for details of the calculation of 
this standard-error performance metric. 
}
\section{Methods}{
\describe{
\item{\code{signature(betas = "exprmethy450")}}{
 all of the methods simply extract betas from the data object 
(which can be a \code{exprmethy450}, \code{MethylSet}, 
\code{MethyLumiSe}, or \code{RGChannelSet}) and calculate 
the metric.
}

}

}
\keyword{methods}
\keyword{ ~~ other possible keyword(s) ~~ }
