% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEPIC.R
\name{NChannelSetToMethyLumiSet2}
\alias{NChannelSetToMethyLumiSet2}
\title{For internal use, is read using minfi-like machinery and then
  preprocessed into the more flexible and convenient methylumi 
  object used by wateRmelon/bigmelon}
\usage{
NChannelSetToMethyLumiSet2(
  NChannelSet,
  parallel = F,
  pval = 0.05,
  n = F,
  n.sd = F,
  oob = T,
  to = TRUE
)
}
\arguments{
\item{NChannelSet}{an NChannelSet (raw red and green values not yet 
mapped to Illumina IDs/CpG names}

\item{parallel}{no effect, included for future parallelisation}

\item{pval}{detection pval threshold for filtering.  Inactivated.}

\item{n}{keep nbeads data (min of m & u)}

\item{n.sd}{process SD of U and M (not currently implemented)}

\item{oob}{keep out-of-band signals}

\item{to}{does chip have type I and II probes?}
}
\value{
A methylumi object with betas, U and M, optionally 
  additional data
}
\description{
For internal use, is read using minfi-like machinery and then
  preprocessed into the more flexible and convenient methylumi 
  object used by wateRmelon/bigmelon
}
