% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifesto.R
\name{read.manifest}
\alias{read.manifest}
\title{read.manifest - read in csv format Illumina chip manifest files}
\usage{
read.manifest(file)
}
\arguments{
\item{file}{}
}
\value{
a list of of dataframes of data prepared for making 
  IlluminaMethylationManifest
}
\description{
read.manifest - read in csv format Illumina chip manifest files
}
\details{
This function is probably not much use for calling directly. 
  It mostly exists to be called by canno.
}
