\name{wateRmelon-package}
\alias{wateRmelon-package}
\alias{wateRmelon}
\docType{package}
\title{
Illumina 450K arrays: normalization and performance metrics
}
\description{
Functions for calculating the index of DNA methylation proportion beta in 15 different ways, and three different ways of estimating data quality or normalization performance. 
}
\details{
\tabular{ll}{
Package: \tab wateRmelon\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-10-10\cr
License: \tab GPL3\cr
}
}
\author{
Leonard C Schalkwyk, Ruth Pidsley and Chloe Wong
Maintainer: Who to complain to <leonard.schalkwyk@kcl.ac.uk>

}
\references{
Pidsley R, Wong CCY, Volta M, Lunnon K, Mill J, Schalkwyk LC: 
A data-driven approach to preprocessing Illumina 450K methylation 
array data (submitted)
}
\keyword{ package }
%\seealso{
%}

