% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAny.R
\name{readPepo}
\alias{readPepo}
\title{readPepo - (unfinished) read any kind of Illumina DNA methylation array idat files into a methylumi object}
\usage{
readPepo(
  idatdir = ".",
  filelist = NULL,
  barcodelist = NULL,
  manifest = NULL,
  parallel = F,
  n = F,
  pdat = NULL,
  oob = F,
  two = TRUE
)
}
\arguments{
\item{idatdir}{the directory with the idatfiles. Currently only handle one directory.}

\item{filelist}{optional list of idat files to process.}

\item{barcodelist}{optional list of barcodes to process.}

\item{manifest}{name of a IlluminaMethylationManifest object or a csv format manifest.  If missing,  will run idet() on one of the idat files.}

\item{parallel}{try to use multiple cores.}

\item{n}{keep beadcounts.}

\item{pdat}{optional data.frame describing the samples.}

\item{two}{are there two different assay types (true of human methylation arrays except 27k)}

\item{keep}{out-of-band (OOB) or opposite-channel signals}
}
\value{
A ‘MethyLumiSet’ object.
}
\description{
readPepo - (unfinished) read any kind of Illumina DNA methylation array idat files into a methylumi object
}
